requirements.txt
asgiref==3.8.1
astroid==3.2.2
boto3==1.34.145
botocore==1.34.145
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cssbeautifier==1.15.1
dill==0.3.8
Django==5.0.6
django-crispy-forms==2.2
django-storages==1.14.4
djlint==1.34.1
EditorConfig==0.12.4
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
isort==5.13.2
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
mccabe==0.7.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pylint==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2023.12.25
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
tomlkit==0.13.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
